requirements.txt
flask_wtf>=0.14.2
flask_socketio>=4.3.0
flask_login>=0.3.0
flask_mail>=0.9.1
flask_pymodm>=0.1.2
flask_bootstrap>=3.3.7.1
flask_classy>=0.6.10
python-dateutil>=2.1
gevent>=20.5.2
gevent-websocket>=0.10.1
git+git://github.com/fastogt/pyfastocloud@master
git+git://github.com/fastogt/pyfastocloud_models@master
git+git://github.com/fastogt/pymodm@master
